backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
django-cors-headers==4.6.0
webauthn>=2.7.0

